+++ src/test/kotlin/blackjack/domain/winning/WinningStatTest.kt
@@ -6,6 +6,7 @@ import blackjack.domain.card.RandomCardDeck.Companion.DIAMOND
 import blackjack.domain.player.Dealer
 import blackjack.domain.player.Player
 import blackjack.domain.score.PlayerScore
+import blackjack.util.winningStat.createWinningStat
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 
@@ -14,7 +15,7 @@ class WinningStatTest : FreeSpec({
     "playerGameResult" - {
 
         "플레이어가 딜러한테 진 경우 게임 결과에 PLAYER_LOOSE이 포함된다." {
-            val winningStat = createWinningStat(2, 3)
+            val winningStat = createWinningStat(1000, 2, 3)
 
             val result = winningStat.playerGameResult()
             result.size shouldBe 1
@@ -23,7 +24,7 @@ class WinningStatTest : FreeSpec({
         }
 
         "플레이어가 딜러한테 이긴 경우 게임 결과에 PLAYER_WIN이 포함된다." {
-            val winningStat = createWinningStat(2, 1)
+            val winningStat = createWinningStat(1000, 2, 1)
 
             val result = winningStat.playerGameResult()
             result.size shouldBe 1
@@ -32,7 +33,7 @@ class WinningStatTest : FreeSpec({
         }
 
         "딜러가 21점을 초과한 경우 DEALER_BUST가 포함된다." {
-            val winningStat = createWinningStat(2, 22)
+            val winningStat = createWinningStat(1000, 2, 22)
 
             val result = winningStat.playerGameResult()
             result.size shouldBe 1
@@ -41,7 +42,7 @@ class WinningStatTest : FreeSpec({
         }
 
         "플레이어가 21점을 초과한 경우 PLAYER_BUST가 포함된다." {
-            val winningStat = createWinningStat(22, 20)
+            val winningStat = createWinningStat(1000, 22, 20)
 
             val result = winningStat.playerGameResult()
             result.size shouldBe 1
@@ -50,7 +51,7 @@ class WinningStatTest : FreeSpec({
         }
 
         "플레이어와 딜러가 동일한 점수일 경우 PLAYER_TIE가 포함된다." {
-            val winningStat = createWinningStat(20, 20)
+            val winningStat = createWinningStat(1000, 20, 20)
 
             val result = winningStat.playerGameResult()
             result.size shouldBe 1
@@ -74,11 +75,3 @@ class WinningStatTest : FreeSpec({
         }
     }
 })
-
-fun createWinningStat(playerScore: Int, dealerScore: Int): WinningStat {
-    val player = Player("test", 1000)
-    return WinningStat(
-        listOf(PlayerScore(player, playerScore)),
-        PlayerScore(Dealer(), dealerScore)
-    )
-}
